# doc/api.yaml
openapi: 3.0.1
info:
  title: WasaPhoto
  description: |
    API specification for WASAPhoto, University project
    for web and software architecture course.
  version: 1.0.2
servers:
  - url: http://localhost:3000
...
paths:
  /session:
    post:
      security:
        - bearerAuth: []
      tags:
        - Login
      summary: logs in the  user
      description: |-
        if the user does not exist it will be created
        else it will be logged in; either way the userId is returned
      operationId: doLogin
      requestBody:
        description: Username to be used for loggin in
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserName"

        required: true

      responses:
        "201":
          description: user log in action succesful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

        "403":
          $ref: "#/components/responses/Forbidden"

        "500":
          $ref: "#/components/responses/InternalServerError"
  /users:
    get:
      security:
      - bearerAuth: []
      tags:
        - User
      summary: retrieves all users
      description: returns a list of all users registered on the platform
      operationId: getUserList
      responses:
        "200":
          description: list of all users retrieved succesfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserFromQuery"
                minItems: 0

        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
    
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/{userId}/profile:
    get:
      security:
        - bearerAuth: []
      tags:
        - User
        # consider putting this under profile tag, or at least # # add 2 tags
      summary: gets user profile
      description: returns the profile of the user or an error if the user does not exist
      operationId: getUserProfile
      parameters:
        - $ref: "#/components/parameters/userId"

      responses:
        "200":
          description: user profile retrieved succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden" # if banned or no auth
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/{userId}/username:
    description: username specific field of the user, can be set or updated
    parameters:
      - $ref: "#/components/parameters/userId"

    put:
      security:
        - bearerAuth: []
      tags:
        - User
      summary: change the username
      description: used to update the userneme or setting it for the current user logged in
      operationId: setMyUserName
      requestBody:
        description: a username to be set
        required: true
        content:
          text/plain:
            schema:
              $ref: "#/components/schemas/UserName"

      responses:
        "204":
          description: username set succesfully

        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/{userId}/stream:
    parameters:
      - $ref: "#/components/parameters/userId"

    get:
      security:
        - bearerAuth: []
      tags:
        - User
      summary: retrieves the stream of a user
      description: |-
        returns personal stream of photos for the user 
        parameters
      operationId: getMyStream
      responses:
        "200":
          description: user stream retrieved successfully
          content:
            application/json:
              schema:
                type: array
                description: list of photos in the stream 
                items:
                  $ref: "#/components/schemas/Photo"
                minItems: 0
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"

        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
...
    BadRequest:
      description: syntax of the request was not compliant with docs, check parameters
    #403
    Forbidden:
      description: server will not accept request due to lacking of client permissions
    #404
    NotFound:
      description: resource not found

    #500 class responses
    #500
    InternalServerError:
      description: the server encountered an error while processing the request
  parameters:
    userId:
      description: identifier for a single user
      schema:
        $ref: "#/components/schemas/Identifier"
      name: userId
      in: path
      required: true
...
  schemas:
    Identifier:
      title: identifier
      description: alphanumerical identifier for a resource
      type: string
      readOnly: true
      #consider specializing it
      pattern: "^@[A-Z][0-9]{9}$"
      example: "@U000000001"
      minLength: 11
      maxLength: 11
    UserName:
      title: username
      description: user name associated with userId, part of a user obj
      type: string
      example: neosnakex34
      pattern: "^[a-z0-9]*?$"
      minLength: 3
      maxLength: 12
    UserFromQuery:
      title: userFromQuery
      description: resulting user from a query in a serchbox 
      type: object
      properties: 
        userId:
          $ref: "#/components/schemas/Identifier"
        username:
          $ref: "#/components/schemas/UserName"
        isRequestorBanned:
          description: boolean value that tells if user is banned by the requestor
          type: boolean
          example: false
        requestorHasBanned:
          description: boolean value that tells if requestor has banned the user
          type: boolean
          example: true 
        requestorHasFollowed:
          description: boolean value that tells if requestor has followed the user
          type: boolean
          example: false
    User:
      title: user
      description: user of WasaPhoto can do a number of actions
      type: object
      properties:
        userId:
          $ref: "#/components/schemas/Identifier"
          #description: unique identifier for a single user

        username:
          $ref: "#/components/schemas/UserName"
    UserProfile:
      title: userProfile
      description: a user profile, providing info about follwers and following, photos and user related stuff
      type: object
      properties:
        userId:
          $ref: "#/components/schemas/Identifier"
        username:
          $ref: "#/components/schemas/UserName"

        followerCounter:
          description: number of users following the owner of the profile
          type: integer
          minimum: 0
          example: 50

        followingCounter:
          description: number of users followed by the owner of the profile
          type: integer
          minimum: 0
          example: 12

        photoCounter:
          description: number of photos uploaded by the owner of the profile
          type: integer
          minimum: 0
          example: 336

        photos:
          description: list of photos uploaded by the user
          type: array
          minItems: 0
          items:
            $ref: "#/components/schemas/Photo"

    # in backend i will take file, convert into bytes and use it through the program
    # i can be more RESTful with those description, try to fix later
    PhotoFile:
      title: photoFile
      description: the actual file of the photo, as a stream of bytes retrieved by it's path in filesystem
      properties:
        path:
          description: position in filesystem for a file
          type: string
          example: "/tmp/wasaphoto/photofiles/rei.jpg"

        file:
          description: the actual file of the photo
          type: string
          format: binary
          minLength: 8 # to fit check for pngs header
          maxLength: 10485760 #binary 10 megs (2 power of 20 times 10)
    Photo:
      title: photo
      description: a photo on the platform
      type: object
      properties:
        photoId:
          $ref: "#/components/schemas/Identifier"

        # uploader user id
        uploaderUserId:
          $ref: "#/components/schemas/Identifier"
        
        uploaderUsername:
          $ref: "#/components/schemas/UserName"
        
        likes:
          description: number of likes the photo has recieved (represented by userId that liked the photo)
          type: integer
          example: 50

        likedByLoggedUser:
          description: boolean value that tells if the photo is liked by the logged user will be computed when retrieving photo
          type: boolean
          example: true

        comments:
          description: comments attached to the photo, made by users
          type: array
          minItems: 0
          maxItems: 20 # not for real use, just as an example
          items:
            $ref: "#/components/schemas/Comment"

        date:
          description: date of uploading
          type: string
          format: date-time
          example: "2021-05-12T09:00:00Z"
          minLength: 20
          maxLength: 20

        # what will be a photopath in struct (path + data)
        rawPhotoFile:
          $ref: "#/components/schemas/PhotoFile"
    CommentBody:
      description: body of the comment
      type: string
      minLength: 1
      maxLength: 400
      pattern: "^.*?$"
      example: "nice one bro"
    Comment:
      title: comment
      description: a comment on a photo
      type: object
      properties:
        commentId:
          $ref: "#/components/schemas/Identifier"

        userId:
          $ref: "#/components/schemas/Identifier"

        photoId:
          $ref: "#/components/schemas/Identifier"

        body:
          $ref: "#/components/schemas/CommentBody"

        date:
          description: date of the comment
          type: string
          format: date-time
          example: "2023-12-12T09:00:00Z"
          minLength: 20
          maxLength: 20